+++ app/Main.hs
@@ -2,6 +2,7 @@ module Main (main) where
 
 import Haskellorls
 import System.Environment (getArgs)
+import System.Exit
 
 main :: IO ()
-main = getArgs >>= haskellorls
+main = getArgs >>= haskellorls >>= exitWith